feat: configure docker and crud

+++ src/main/java/com/lotus/igaming/igaming/repositories/ClientRepository.java
@@ -1,9 +1,11 @@
 package com.lotus.igaming.igaming.repositories;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lotus.igaming.igaming.Client;
 
-public interface ClientRepository extends JpaRepository<Client, Integer> {
+public interface ClientRepository extends JpaRepository<Client, Long> {
 
 }
